tutorials/ml-pipeline-with-workflows/services/train/requirements.txt
Flask
GitPython
google_apitools
google-api-python-client
gunicorn
six>=1.13.0,<2dev
